[conftest] move overridden ansible fixture to seperate file
